simplify: detach logits under stop-grad + cleanup ordering-loss plumbing
Review findings addressed:
1. Critical — CE_STOP_GRAD/GRADIENT_STOP gradient leak. The ordering loss was
   using raw classifier_logits_out, defeating the stop-grad contract that
   isolates the classifier from regression-path gradients. Now detached at the
   same point as the MDN path (line 167).
2. Removed redundant 'or 0.0' and float(...) casts around getattr reads in
   _calculate_custom_loss. _defaults already supplies the floats.
3. Dropped redundant int(math.ceil(...)) in ordering_loss.py — math.ceil
   returns int in Python 3.
4. Extracted the common ProbReg fit() setup in the integration tests into a
   module-scoped fixture, halving per-test .fit() cost.

Added two gradient-contract tests:
- test_probreg_ordering_under_ce_stop_grad_does_not_leak_to_classifier
- test_probreg_ordering_under_regression_only_does_flow_to_classifier

Regression: 48/48 tests green across probreg + phase1 + new files.

+++ automl_package/utils/ordering_loss.py
@@ -42,7 +42,7 @@ def compute_ordering_means(
     batch_size, k = classifier_logits.shape
     probs = torch.softmax(classifier_logits, dim=-1)  # (B, k)
 
-    subset_size = max(1, int(math.ceil(batch_size * top_decile_fraction)))
+    subset_size = max(1, math.ceil(batch_size * top_decile_fraction))
 
     # For each head i, find indices of the top subset_size samples by p_i.
     # topk over dim=0 picks the B highest along batch for each column.
